move writeBasedOnSchema to IndexUtils
